move get_enriched_config into config provider
